+++ axum-extra/src/extract/cookie.rs
@@ -1,6 +1,6 @@
 //! Cookie parsing and cookie jar management.
 //!
-//! See [`CookieJar`] and [`SignedCookieJar`] for more details.
+//! See [`CookieJar`], [`SignedCookieJar`], and [`PrivateCookieJar`] for more details.
 
 use axum::{
     async_trait,
@@ -15,7 +15,10 @@ use http::{
 };
 use std::{convert::Infallible, fmt, marker::PhantomData};
 
-pub use cookie_lib::{Cookie, Key};
+mod private;
+
+pub use self::private::PrivateCookieJar;
+pub use cookie_lib::{Cookie, Expiration, Key, SameSite};
 
 /// Extractor that grabs cookies from the request and manages the jar.
 ///
@@ -273,7 +276,7 @@ where
     type Rejection = <axum::Extension<K> as FromRequest<B>>::Rejection;
 
     async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        let Extension(key) = Extension::from_request(req).await?;
+        let key = Extension::<K>::from_request(req).await?.0.into();
 
         let mut jar = cookie_lib::CookieJar::new();
         let mut signed_jar = jar.signed_mut(&key);
@@ -377,7 +380,7 @@ impl<K> SignedCookieJar<K> {
     }
 }
 
-impl IntoResponseParts for SignedCookieJar {
+impl<K> IntoResponseParts for SignedCookieJar<K> {
     type Error = Infallible;
 
     fn into_response_parts(self, mut res: ResponseParts) -> Result<ResponseParts, Self::Error> {
@@ -397,7 +400,7 @@ fn set_cookies(jar: cookie_lib::CookieJar, headers: &mut HeaderMap) {
     // jar so it cannot be called multiple times.
 }
 
-impl IntoResponse for SignedCookieJar {
+impl<K> IntoResponse for SignedCookieJar<K> {
     fn into_response(self) -> Response {
         (self, ()).into_response()
     }
@@ -448,7 +451,8 @@ mod tests {
                     .route("/set", get(set_cookie))
                     .route("/get", get(get_cookie))
                     .route("/remove", get(remove_cookie))
-                    .layer(Extension(Key::generate()));
+                    .layer(Extension(Key::generate()))
+                    .layer(Extension(CustomKey(Key::generate())));
 
                 let res = app
                     .clone()
@@ -492,6 +496,18 @@ mod tests {
 
     cookie_test!(plaintext_cookies, CookieJar);
     cookie_test!(signed_cookies, SignedCookieJar);
+    cookie_test!(signed_cookies_with_custom_key, SignedCookieJar<CustomKey>);
+    cookie_test!(private_cookies, PrivateCookieJar);
+    cookie_test!(private_cookies_with_custom_key, PrivateCookieJar<CustomKey>);
+
+    #[derive(Clone)]
+    struct CustomKey(Key);
+
+    impl From<CustomKey> for Key {
+        fn from(custom: CustomKey) -> Self {
+            custom.0
+        }
+    }
 
     #[tokio::test]
     async fn signed_cannot_access_invalid_cookies() {