+++ axum/src/extract/query.rs
@@ -37,8 +37,8 @@ use std::ops::Deref;
 /// # };
 /// ```
 ///
-/// If the query string cannot be parsed it will reject the request with a `400
-/// Bad Request` response.
+/// If the query string cannot be parsed it will reject the request with a `422
+/// Unprocessable Entity` response.
 ///
 /// For handling values being empty vs missing see the (query-params-with-empty-strings)[example]
 /// example.